Re-factoring and ldy
* Refactor: Added comments to global variables and certain functions.
Changed the instruction function pointer type to include a parameter:
the opcode. This is so that the fetch_operand() function can deduce
what the size of the instruction is. A page_boundary_crossed() function
was added. This will be used by instructions where there is a
possibility of a page being crossed. All lda, ldx, adci, sta
instructions have been refactored to reflect these (fetch_operand() and
page_boundary_crossed()) changes.

* Added ldy family of instructions.